+++ src/app/builder/edit/[buildId]/layout.tsx
@@ -1,4 +1,5 @@
 import { isErrorResponse } from '@/features/error-handling/isErrorResponse'
+import { PageHeader } from '@/features/ui/PageHeader'
 
 import { getBuild } from '../../actions'
 import BuildPage from './page'
@@ -11,9 +12,14 @@ export default async function Layout({
   const buildData = await getBuild(buildId)
   if (isErrorResponse(buildData)) {
     console.error(buildData.errors)
-    throw new Error(`Build ${buildId} is not found. If you are sure the build exists, it may
-    be marked private. You must be logged in as the build creator to view
-    a private build.`)
+    return (
+      <div className="flex max-w-lg flex-col">
+        <PageHeader
+          title="Something went wrong!"
+          subtitle="The build either can't be found or is marked private."
+        />
+      </div>
+    )
   }
   const { build } = buildData