+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Added
 
 - ConRaven the God and BrotherFromHood have added even more aberration images.
+- Delete Build button now available when viewing a build or editing a build.
 
 ### Changed